Refactor deposit to accept deposit data instead of targeting only erc20 deposits
